Fixes #2736. Adds BaseSplitDetailView to avoid duplicated code. Adds Android layout for SplitDetailNavView.
