+++ Projects/Playground/Playground.Droid/Views/Base/BaseSplitDetailView.cs
@@ -1,0 +1,64 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MS-PL license.
+// See the LICENSE file in the project root for more information.
+
+using Android.Content.Res;
+using Android.OS;
+using Android.Support.V7.Widget;
+using Android.Views;
+using MvvmCross.Droid.Support.V4;
+using MvvmCross.Droid.Support.V7.AppCompat;
+using MvvmCross.Platforms.Android.Binding.BindingContext;
+using MvvmCross.ViewModels;
+
+namespace Playground.Droid.Views.Base
+{
+    public abstract class BaseSplitDetailView<TViewModel> : MvxFragment<TViewModel> where TViewModel : class, IMvxViewModel
+    {
+        protected SplitRootView BaseActivity => (SplitRootView)Activity;
+        protected Toolbar _toolbar;
+        protected MvxActionBarDrawerToggle _drawerToggle;
+
+        protected abstract int FragmentLayoutId { get; }
+
+        public override View OnCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
+        {
+            base.OnCreateView(inflater, container, savedInstanceState);
+
+            var view = this.BindingInflate(FragmentLayoutId, null);
+
+            _toolbar = view.FindViewById<Toolbar>(Resource.Id.toolbar);
+            if (_toolbar != null)
+            {
+                BaseActivity.SetSupportActionBar(_toolbar);
+                BaseActivity.SupportActionBar.SetDisplayHomeAsUpEnabled(true);
+
+                _drawerToggle = new MvxActionBarDrawerToggle(
+                    Activity,                               // host Activity
+                    BaseActivity.DrawerLayout,              // DrawerLayout object
+                    _toolbar,                               // nav drawer icon to replace 'Up' caret
+                    Resource.String.drawer_open,            // "open drawer" description
+                    Resource.String.drawer_close            // "close drawer" description
+                );
+                BaseActivity.DrawerLayout.AddDrawerListener(_drawerToggle);
+                _drawerToggle.DrawerIndicatorEnabled = true;
+            }
+
+            return view;
+        }
+
+        public override void OnConfigurationChanged(Configuration newConfig)
+        {
+            base.OnConfigurationChanged(newConfig);
+            if (_toolbar != null)
+                _drawerToggle.OnConfigurationChanged(newConfig);
+        }
+
+        public override void OnActivityCreated(Bundle savedInstanceState)
+        {
+            base.OnActivityCreated(savedInstanceState);
+            if (_toolbar != null)
+                _drawerToggle.SyncState();
+        }
+    }
+}
